pyproject.toml
[tool.poetry]
name = "amc-bc-protein-digger"
version = "0.1.0"
description = ""
authors = ["Jinahn J <jinnannann@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
pyyaml = "^6.0.2"
pathlib = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
